Implemented sound system and fixed Movement bugs
I added a sound system that plays footsteps when the play walks on the ground and stops when the player jumps. I also fixed a bug where my movement wasn't Camera-Relative and speed inconsistencies. Also added a texture to the floor to help with debugging but it also may contribute to one of the skill/task list