Imported Asset and (transparent) textures done
Imported an asset of the web and applied textures. Although the idea was to make the lenses transparent but the texture maps didn't have an alpha map and I couldnt figure out how to do it to specific parts. So I ended up just making the entire asset a little bit transparent