PR_02 Updates
When player overs over interactable items I made told them to "Press 'E' to..." so they don't have to figure out what the interact button is. I also Made the crosshair and text colour different so it can be see better. Also move the text up a little bit when it appears.